youtube_timeline_generator/requirements.txt
# YouTube動画処理
yt-dlp>=2024.1.0

# 音声認識
openai-whisper>=20231117

# 音声処理
pydub>=0.25.1

# 自然言語処理
nltk>=3.8
scikit-learn>=1.3.0
numpy>=1.24.0

# ユーティリティ
tqdm>=4.66.0
python-dotenv>=1.0.0
click>=8.1.7

# ログ
loguru>=0.7.2

# エンタメモード用（音声解析）
librosa>=0.10.0
scipy>=1.11.0
matplotlib>=3.7.0
soundfile>=0.12.0